@@ -1,14 +1,15 @@
 {
   "name": "chase-center-calendar",
+  "type": "module",
   "scripts": {
-    "eslint": "eslint .",
+    "eslint": "eslint -c .eslint.config.js .",
     "stylelint": "stylelint static/css/*",
-    "test": "npm run build:dev && npm run eslint && npm run stylelint && npm run cover",
-    "cover": "nyc --require dotenv/config --instrumenter false mochify --global-transform browserify-css --transform loose-envify --require source-map-support/register --transform [ babelify --ignore [ test ] --plugins [ babel-plugin-istanbul ] ] static/js/test --allow-chrome-as-root",
+    "test": "npm run build:dev && npm run eslint && npm run stylelint && npm run wdio",
     "build": "webpack --mode=production --node-env=production",
     "build:dev": "webpack --mode=development",
     "build:prod": "webpack --mode=production --node-env=production",
-    "watch": "webpack --watch"
+    "watch": "webpack --watch",
+    "wdio": "wdio run ./wdio.conf.js"
   },
   "repository": {
     "type": "git",
@@ -18,14 +19,11 @@
   "private": true,
   "dependencies": {
     "@popperjs/core": "^2.11.7",
-    "babelify": "^10.0.0",
     "bootstrap": "^5.0.0",
     "css-loader": "^6.7.3",
     "css-minimizer-webpack-plugin": "^6.0.0",
-    "dotenv": "^16.0.0",
     "dotenv-webpack": "^8.0.1",
-    "logfit": "^0.8.2",
-    "loose-envify": "^1.4.0",
+    "logfit": "^0.9.0",
     "mini-css-extract-plugin": "^2.7.5",
     "normalize.css": "^8.0.1",
     "rollbar": "^2.21.0",
@@ -34,13 +32,14 @@
     "webpack-cli": "^5.0.2"
   },
   "devDependencies": {
-    "babel-plugin-istanbul": "^6.0.0",
-    "browserify-css": "^0.15.0",
-    "chai": "^4.3.0",
-    "eslint": "^8.0.1",
-    "mochify": "^9.0.0",
-    "nyc": "^15.1.0",
-    "source-map-support": "^0.5.19",
+    "@wdio/browser-runner": "^8.35.1",
+    "@wdio/cli": "^8.35.1",
+    "@wdio/mocha-framework": "^8.35.0",
+    "@wdio/spec-reporter": "^8.32.4",
+    "chai": "^5.0.0",
+    "dotenv": "^16.0.0",
+    "eslint": "^9.0.0",
+    "globals": "^15.0.0",
     "stylelint": "^16.0.0",
     "stylelint-config-standard": "^36.0.0"
   },
